refactor & handle error
