+++ src/events/channel_delete.ts
@@ -1,7 +1,9 @@
 import {Client} from "discord.js";
 import fs from "fs";
 
-export default (client	: Client): void => {
+import { GuildData } from "../interface";
+
+export const channel_delete = (client	: Client): void => {
 	client.on("channelDelete", async (channel) => {
 		try {
 			if (channel.isDMBased()) return;
@@ -19,5 +21,48 @@ export default (client	: Client): void => {
 			console.error(error);
 		}
 		
+	});
+};
+
+export const delete_message = (client	: Client): void => {
+	client.on("messageDelete", async (message) => {
+		try {
+			if (!message.guild) return;
+			const messageId = message.id;
+			//search channelID in database and delete it
+			const guildID = message.guild.id;
+			const database = fs.readFileSync("database.json", "utf-8");
+			const parsedDatabase = JSON.parse(database);
+			if (!parsedDatabase[guildID]) return;
+			const guildData = parsedDatabase[guildID] as Partial<GuildData>;
+			if (guildData?.templateID?.messageId === messageId) {
+				delete guildData.templateID;
+			}
+			const dbUser = guildData?.user;
+			if (dbUser && Object.keys(dbUser).length > 0){
+				for (const values of Object.values(dbUser)) {
+					if (values.length === 0) continue;
+					for (const [index, value] of values.entries()) {
+						if (value.messageId === messageId) {
+							values.splice(index, 1);
+						}
+					}
+				}
+			}
+			fs.writeFileSync("database.json", JSON.stringify(parsedDatabase, null, 2), "utf-8");
+		} catch (error) {
+			console.error(error);
+		}
+	});
+};
+
+export const on_kick = (client: Client): void => {
+	client.on("guildDelete", async (guild) => {
+		//delete guild from database
+		const guildID = guild.id;
+		const data = fs.readFileSync("database.json", "utf-8");
+		const json = JSON.parse(data);
+		if (json[guildID]) delete json[guildID];
+		fs.writeFileSync("database.json", JSON.stringify(json, null, 2));
 	});
 };
